requirements.txt
asgiref==3.7.2
cbor2==5.6.0
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
cryptography==42.0.0
defusedxml==0.8.0rc2
Django==5.0.1
django-cors-headers==4.3.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
-e git+https://github.com/sunscrapers/djoser.git@4452009f628b59f02704373d5e7f991f1243397f#egg=djoser
drf-yasg==1.21.7
future==0.18.3
idna==3.6
inflection==0.5.1
oauthlib==3.2.2
packaging==23.2
pillow==10.2.0
pycountry==23.12.11
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==24.0.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
reportlab==4.1.0
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.1
sqlparse==0.4.4
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.1.0
webauthn==0.4.7
